Fix the navigation component's active class handling
The active class had a typo and once corrected, the component wouldn't behave as expected because the active class would be set by the router and removed with the `[class.-active]` directive. To fix this, the class is handled entirely by a custom method that determine if the route is active.