Fix the pagination issue
The table would indicate more pages than available. The issue was that the filters would be set after the table's data would be request so if a default filter would be set, only subsequent requests to get a page would get the correct page count.

In addition, the weight of the queries to get the filters values has been reduced drastically by only requesting the necessary fields.